Add option handling to registry/parse.c
This commit updates src/registry/parse.c to contain logic for passing code
generation options to the registry parser. By default, parse will read from
standard in. An option can be passed to parse by passing a string that contains
`--` to argv. This will trigger the command argument to be processed as an
option rather then a file. All other arguments (that don't start with `--`)
will be opened as a file.