Update landuse and soilcat to use refactored mapping function
This commit updates landuse (lu_index) and soil category top (soilcat_top) to
use the unified mapping function that was introduced in c117d623b. It also
moves the landmask derivation, and soiltemp interpolation into their own
functions outside of init_atm_static.

The interpolation of soiltemp does not need to use the unified mapping function
as it can be safely interpolated in parallel.